requirements.txt
absl-py==1.0.0
alabaster==0.7.12
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
autograd==1.4
Babel==2.10.1
backcall==0.2.0
basemap==1.3.3
basemap-data==1.3.2
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.2.0
colorama==0.4.5
commonmark==0.9.1
cuda-python==11.7.1
cycler==0.11.0
Cython==0.29.32
decorator==5.1.1
distlib==0.3.4
dm-tree==0.1.7
docopt==0.6.2
docutils==0.17.1
earthengine-api==0.1.316
entrypoints==0.4
executing==0.8.3
filelock==3.6.0
Fiona==1.8.21
firelight==0.2.1
flatbuffers==2.0.7
fonttools==4.33.3
functorch==0.1.1
future==0.18.2
gast==0.4.0
geomstats==2.5.0
geopandas==0.11.0
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.8.2
google-api-python-client==1.12.11
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.1
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.3
grpcio==1.46.1
h5py==3.7.0
httplib2==0.20.4
httplib2shim==0.0.3
idna==3.3
imageio==2.19.3
imagesize==1.3.0
importlib-metadata==4.11.3
iniconfig==1.1.1
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonpickle==0.9.6
jupyter-client==7.3.5
jupyter-core==4.11.1
keopscore==2.1
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==14.0.6
llvmlite==0.39.1
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
munch==2.5.0
nest-asyncio==1.5.6
networkx==2.8.4
numba==0.56.2
numpy==1.22.3
oauthlib==3.2.0
openTSNE==0.6.2
opt-einsum==3.3.0
ordereddict==1.1
packaging==21.3
pandas==1.4.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==2.0.10
protobuf==3.19.5
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py-cpuinfo==8.0.0
pyaml==21.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.1
pygeos==0.12.0
Pygments==2.12.0
pykeops==2.1
pynndescent==0.5.7
pyparsing==3.0.8
pyproj==3.3.1
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
pyreadr==0.4.6
pyshp==2.1.3
pytest==7.1.2
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==24.0.1
qc-procrustes==1.0.0
requests==2.27.1
requests-oauthlib==1.3.1
rich==12.4.4
rsa==4.8
sacred==0.7.4
scikit-image==0.19.3
scikit-learn==1.1.2
scikit-optimize==0.5.2
scipy==1.8.0
Shapely==1.8.2
six==1.16.0
sklearn==0.0
smmap==5.0.0
snowballstemmer==2.2.0
Sphinx==4.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.2.0
tadasets==0.0.4
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
tensorflow-probability==0.18.0
termcolor==2.0.1
threadpoolctl==3.1.0
tifffile==2022.5.4
tomli==2.0.1
torch==1.11.0
torchaudio==0.11.0
torchdiffeq==0.2.3
torchvision==0.12.0
tornado==6.2
tqdm==4.64.0
traitlets==5.2.1.post0
typing-extensions==4.1.1
umap-learn==0.5.3
uritemplate==3.0.1
urllib3==1.26.9
virtualenv==20.14.1
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
